+++ tests/test_main.py
@@ -1,10 +1,10 @@
 from pytest_mock import MockerFixture
 from typer.testing import CliRunner
 
+from pyrb.brokerage.context import RebalanceContext
+from pyrb.brokerage.order_manager import Order, OrderType
 from pyrb.controller import app
 from pyrb.exceptions import InsufficientFundsException
-from pyrb.repositories.context import RebalanceContext
-from pyrb.repositories.order_manager import Order, OrderType
 
 
 def test_sut_rebalances(fake_rebalance_context: RebalanceContext, mocker: MockerFixture) -> None:
